+++ test/cases/signin-provider-alias-type-inference/nuxt.config.ts
@@ -1,0 +1,10 @@
+import { fileURLToPath } from 'node:url'
+
+const serverConfig = fileURLToPath(new URL('./server/auth.config', import.meta.url))
+
+export default defineNuxtConfig({
+  extends: ['../_base-module'],
+  auth: {
+    serverConfig,
+  },
+})